Rename: --imports to --symbols
