+++ src/cli/symbols.rs
